Fixed sorting by columns other than the filename.  We were missing a cast to unsigned short when using MAKE_HRESULT.  Why this cast is so crucial, I don't know but it is so leave it there!

+++ Folder.cpp
@@ -268,7 +268,9 @@ STDMETHODIMP CFolder::CompareIDs(
 
 		if ((::ILIsChild(pidl1) && ::ILIsChild(pidl2)) || result != 0)
 		{
-			return MAKE_HRESULT(SEVERITY_SUCCESS, 0, result);
+			// The cast to unsigned short is *crucial*!  Without it, sorting
+			// in Explorer does all sorts of wierd stuff
+			return MAKE_HRESULT(SEVERITY_SUCCESS, 0, (unsigned short)result);
 		}
 		else
 		{
